Take details out of impl
